requirements.txt
selenium
scikit-learn
numpy
textblob
python-Levenshtein
pillow
